+++ Cargo.toml
@@ -15,9 +15,10 @@ nix = { version = "0.30", features = [ "socket", "uio", "net", "fs" ] }
 serde = { version = "1.0", features = [ "derive" ] }
 serde_json = "1.0"
 sd-notify = "0.4"
-libc = "0.2"
 const_format = "0.2"
 arc-swap = "1.8"
+dashmap = "6.1"
+moka = { version = "0.12", features = [ "future" ] }
 
 [build-dependencies]
 bindgen = "0.72"
