+++ src/handlers/mod.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 pub(super) mod constants;
-pub(super) mod forwarders;
-pub(self) mod helpers;
 pub(super) mod signal_handler;
+pub(super) mod tcp_forwarder;
+pub(super) mod udp_forwarder;
