+++ src/utils/constants.rs
@@ -5,7 +5,7 @@ use const_format::concatcp;
 use super::cap_bindings::{__user_cap_header_struct, CAP_NET_ADMIN, CAP_NET_BIND_SERVICE};
 
 /// Current PID
-pub(crate) static PID: LazyLock<u32> = LazyLock::new(process::id);
+pub(in super::super) static PID: LazyLock<u32> = LazyLock::new(process::id);
 
 /// Metadata header [`__user_cap_header_struct`] to fetch process capabilities
 pub(super) static CAP_HEADER: LazyLock<__user_cap_header_struct> = LazyLock::new(Default::default);
