+++ src/utils/utils.rs
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use libc::{SYS_capget, syscall};
 use log::LevelFilter;
 use log4rs::{
     Handle,
@@ -10,11 +9,9 @@ use log4rs::{
     filter::threshold::ThresholdFilter,
     init_config,
 };
+use nix::libc::{SYS_capget, syscall};
 use serde_json::from_str;
-use std::{
-    io::{Error, ErrorKind, Result as IoResult},
-    path::PathBuf,
-};
+use std::{io, path::PathBuf};
 use tokio::fs::read_to_string;
 
 use super::{
@@ -29,30 +26,31 @@ use super::{
 /// * Each field of each [`__user_cap_data_struct`] holds 32 of them as u32 bitmap (Hence, two are used)
 /// * When enabled, the corresponding bit in that field is 1
 /// * Here we are using [`__user_cap_data_struct::effective`] for our purpose
-pub(crate) fn is_capable() -> IoResult<bool> {
+pub(in super::super) fn is_capable() -> io::Result<bool> {
     let mut data = <[__user_cap_data_struct; 2] as Default>::default();
 
     match unsafe { syscall(SYS_capget, &*CAP_HEADER as *const _, &mut data as *mut _) } {
         0 => Ok(REQUIRED_CAPS
             .iter()
             .all(|&cap| (data[cap_to_index!(cap)].effective & cap_to_mask!(cap)) != 0)),
-        _ => Err(Error::last_os_error()),
+        _ => Err(io::Error::last_os_error()),
     }
 }
 
 /// Read and parse configuration file
-pub(crate) async fn read_config(path: &PathBuf) -> IoResult<Configs> {
+pub(in super::super) async fn read_config(path: &PathBuf) -> io::Result<Configs> {
     if !path.exists() {
-        return Err(Error::new(ErrorKind::NotFound, "Configuration file not found"));
+        return Err(io::Error::new(io::ErrorKind::NotFound, "Configuration file not found"));
     } else if !path.is_file() {
-        return Err(Error::new(ErrorKind::InvalidInput, "Provided configuration path is not a file"));
+        return Err(io::Error::new(io::ErrorKind::InvalidInput, "Provided configuration path is not a file"));
     }
 
-    from_str(&read_to_string(path).await?).map_err(|e| Error::new(ErrorKind::InvalidData, format!("Failed to deserialize configuration file - {e}")))
+    from_str(&read_to_string(path).await?)
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, format!("Failed to deserialize configuration file - {e}")))
 }
 
 /// Enable logging based on provided optional log directory. If provided it logs to file, else falls back to console logging
-pub(crate) fn enable_logging(log_dir: Option<&PathBuf>) -> Result<Handle, LogError> {
+pub(in super::super) fn enable_logging(log_dir: Option<&PathBuf>) -> Result<Handle, LogError> {
     let config = match log_dir {
         Some(dir) => {
             let metadata = dir
@@ -117,7 +115,7 @@ macro_rules! banner {
     };
 }
 
-pub(crate) use banner;
+pub(in super::super) use banner;
 
 #[cfg(test)]
 mod tests {
@@ -139,12 +137,12 @@ mod tests {
         assert!(!file_path_nonexistent.exists());
         let mut result = read_config(&file_path_nonexistent).await;
         assert!(result.is_err());
-        assert_eq!(ErrorKind::NotFound, result.unwrap_err().kind());
+        assert_eq!(io::ErrorKind::NotFound, result.unwrap_err().kind());
 
         // not a file
         result = read_config(&dir_path).await;
         assert!(result.is_err());
-        assert_eq!(ErrorKind::InvalidInput, result.unwrap_err().kind());
+        assert_eq!(io::ErrorKind::InvalidInput, result.unwrap_err().kind());
 
         // using actual file
         let file_path = dir_path.join("config.conf");
@@ -154,7 +152,7 @@ mod tests {
         write(&file_path, b"abcd").await.unwrap();
         result = read_config(&file_path).await;
         assert!(result.is_err());
-        assert_eq!(ErrorKind::InvalidData, result.unwrap_err().kind());
+        assert_eq!(io::ErrorKind::InvalidData, result.unwrap_err().kind());
 
         let conf = json!({
             "port": 8080,
